requirements.txt
dash
dash-bootstrap-components
plotly
gunicorn
pandas
weasyprint
python-dotenv
statsmodels